Add plugin TypeScript pilot
