+++ packages/plugin/.gitignore
@@ -1,2 +1,2 @@
-build
+dist
 node_modules
