+++ packages/plugin/package.json
@@ -2,12 +2,21 @@
   "name": "@postr/plugin",
   "version": "1.0.0",
   "description": "The base syndication plugin",
-  "main": "index.js",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "files": [
+    "dist"
+  ],
   "scripts": {
+    "build": "tsc",
+    "prepare": "tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "engines": {
     "node": ">=24"
+  },
+  "devDependencies": {
+    "typescript": "^5.7.0"
   },
   "homepage": "https://github.com/grantcodes/postr",
   "bugs": "https://github.com/grantcodes/postr/issues",
